Remove hard dependency on CSM and CLM from the suite auth description classes.
This involves a small API change (some internal-only static methods were moved
from ScopeDescription to SuiteRoleMembershipLoader) and an OSGi metadata
change (the suite auth bundle now treats CSM and CLM as optional imports). The
benefit is that the following classes may now be used in a context in which
CSM is not loaded:

* SuiteRole
* SuiteRoleMembership
* ScopeType
* ScopeDescription

The immediate motivation for this change is to aid in testing PSC
authorization plugins. PSC's authorization plugin API uses SuiteRole and
SuiteRoleMembership but is independent of CSM.

SVN-Revision: 650